backend/requirements.txt

flask==2.3.2
waitress==2.1.2
werkzeug==2.3.4
psycopg2-binary==2.9.5
flask-cors==3.0.10
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
PyJWT==2.8.0
pytz
